Basic User Accounts and Other Improvements
User accounts now working, though it is using the local db. Next step will be to make it work from the remote db because the users db should not be sync-ed with every client. Currently, it moves the games created while not signed-in to the user account that is created. When the user signs out, a new client ID is created. There are some rough ends that need to be fixedthough. Game path also been added and working.